src/API/requirements.txt
fastapi==0.114.2
fastapi-cli==0.0.5
psycopg2-binary==2.9.9
pandas==2.2.2
passlib[bcrypt]==1.7.4
SQLAlchemy==2.0.34
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
prometheus-fastapi-instrumentator
scikit-learn